Restructure custom crediential cards
