+++ app/components/CredentialCard/CredentialCard.d.ts
@@ -1,0 +1,5 @@
+import type { CredentialRecordRaw } from '../../model/credential';
+
+export type CredentialCardProps = {
+  rawCredentialRecord: CredentialRecordRaw,
+}
